Trigger nightly builds on fixed nightly branches
Improve the nightly build script so that it targets
a fixed git branch per ngdevkit repo.

If the build script detects that the nightly version
already exists for the latest commit in the ngdevkit
repo, it is now able to trigger a package re-build by
bumping the brew build revision.

Ref dciabrin/ngdevkit#84